Rename kernel_launch_ops to legacy_kernel_launch_ops
