+++ numba_dpex/dpctl_iface/__init__.py
@@ -14,7 +14,7 @@ includes:
 """
 import numba_dpex.dpctl_iface.dpctl_function_types as dpctl_fn_ty
 from numba_dpex.dpctl_iface.dpctl_capi_fn_builder import DpctlCAPIFnBuilder
-from numba_dpex.dpctl_iface.kernel_launch_ops import KernelLaunchOps
+from numba_dpex.dpctl_iface.legacy_kernel_launch_ops import KernelLaunchOps
 
 __all__ = [
     DpctlCAPIFnBuilder,
